+++ examples/react/src/components/CustomCheckout/index.tsx
@@ -57,7 +57,7 @@ export const CustomCheckout = () => {
     creditCardProvider: 'transak' as CreditCardProviders,
     transakConfig: {
       contractId,
-      apiKey: '5911d9ec-46b5-48fa-a755-d59a715ff0cf'
+      apiKey: process.env.REACT_APP_TRANSAK_API_KEY || ''
     },
     onSuccess: (txnHash?: string) => {
       console.log('success!', txnHash)
@@ -90,7 +90,7 @@ export const CustomCheckout = () => {
                 width: '36px'
               }}
             >
-              <Image disableAnimation src={orderSummary.data?.collectibleItem?.collectibleImageUrl} />
+              <Image src={orderSummary.data?.collectibleItem?.collectibleImageUrl} />
             </div>
             <div className="flex flex-col gap-0.5">
               <Text variant="small" color="secondary" fontWeight="medium">
@@ -104,7 +104,7 @@ export const CustomCheckout = () => {
         </div>
         <div className="flex gap-1 flex-col">
           <div className="flex flex-row gap-2 items-center">
-            <NetworkImage disableAnimation chainId={chainId} size="sm" />
+            <NetworkImage chainId={chainId} size="sm" />
             <Text
               color="white"
               variant="large"