+++ packages/checkout/src/views/Checkout/PaymentMethodSelect/PayWithCreditCard/index.tsx
@@ -1,6 +1,6 @@
+import { findSupportedNetwork } from '@0xsequence/connect'
 import { Button, Spinner, Text } from '@0xsequence/design-system'
 import { useClearCachedBalances, useGetContractInfo } from '@0xsequence/hooks'
-import { findSupportedNetwork } from '@0xsequence/connect'
 import { useConnection } from 'wagmi'
 
 import type { CheckoutSettings } from '../../../../contexts/CheckoutModal.js'
@@ -132,7 +132,9 @@ export const PayWithCreditCardTab = ({ skipOnCloseCallback }: PayWithCreditCardT
         Buy directly with your debit or credit card {getProviderName() ? `through ${getProviderName()}` : ''}
       </Text>
       <div className="flex flex-col justify-center items-center gap-2 w-full h-full">
-        <Button className="w-full" shape="square" onClick={payWithSelectedProvider} label="Continue" variant="primary" />
+        <Button className="w-full" shape="square" onClick={payWithSelectedProvider} variant="primary">
+          Continue
+        </Button>
       </div>
     </div>
   )