+++ packages/checkout/src/views/PaymentSelection/OrderSummary/index.tsx
@@ -1,7 +1,7 @@
 import { CollectibleTileImage, formatDisplay } from '@0xsequence/connect'
+import { findSupportedNetwork } from '@0xsequence/connect'
 import { Spinner, Text, TokenImage } from '@0xsequence/design-system'
 import { useGetCoinPrices, useGetContractInfo, useGetTokenMetadata } from '@0xsequence/hooks'
-import { findSupportedNetwork } from '@0xsequence/connect'
 import { formatUnits } from 'viem'
 
 import { useSelectPaymentModal } from '../../../../hooks/useSelectPaymentModal.js'
@@ -21,7 +21,7 @@ export const OrderSummary = () => {
       tokenIDs: tokenIds.some(id => id === '') ? [] : tokenIds
     },
     {
-      disabled: tokenIds.some(id => id === '')
+      enabled: tokenIds.every(id => id !== '')
     }
   )
   const { data: dataCollectionInfo, isLoading: isLoadingCollectionInfo } = useGetContractInfo({