+++ packages/connect/src/components/PolygonBrand/index.ts
@@ -1,0 +1,2 @@
+export { PoweredByPolygon } from './PoweredByPolygon.js'
+export { PolygonLogo } from './PolygonLogo.js'